+++ app/src/main/java/com/college/examples/ui/MainActivity.java
@@ -1,0 +1,32 @@
+package com.college.examples.ui;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.lifecycle.ViewModelProvider;
+
+import com.college.examples.R;
+import com.college.examples.data.MainViewModel;
+import com.college.examples.databinding.ActivityMainBinding;
+
+public class MainActivity extends AppCompatActivity {
+    private ActivityMainBinding binding;
+    private MainViewModel viewModel;
+
+    //same as public static void main(String args[])
+    @Override //app starts here
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        binding = ActivityMainBinding.inflate(getLayoutInflater());
+
+        viewModel = new ViewModelProvider(this).get(MainViewModel.class);
+        //this is the only function call, loads stuff onto screen
+        setContentView(binding.getRoot());
+
+        binding.myImageButton.setOnClickListener(clik -> {
+            binding.myTextView.setText(R.string.img_clicked_msg);
+
+        });
+    }
+}
